moved most of logic into Calendar class
